# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vessel Telemetry API",
    "version": "1.0.0",
    "description": "API for ingesting and retrieving vessel telemetry data from XLSX files"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
    "/ingest/xlsx": {
      "post": {
        "summary": "Ingest XLSX telemetry file",
        "description": "Upload and process an XLSX file containing vessel telemetry data",
        "parameters": [
          {
            "name": "imo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "IMO number of the vessel (preferred identifier)"
          },
          {
            "name": "vessel_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Name of the vessel (fallback if IMO unknown)"
          },
          {
            "name": "period_start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "ISO 8601 timestamp for batch period start"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "XLSX file to upload"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing parameters"
          },
          "409": {
            "description": "File already ingested"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IngestResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ingested", "already_ingested"]
          },
          "upload_id": {"type": "integer", "format": "int64", "nullable": true},
          "vessel_id": {"type": "integer", "format": "int64", "nullable": true},
          "rows_inserted": {
            "type": "object",
            "properties": {
              "engines": {"type": "integer"},
              "fuel": {"type": "integer"},
              "generators": {"type": "integer"},
              "cctv": {"type": "integer"},
              "impact": {"type": "integer"},
              "location": {"type": "integer"}
            }
          },
          "warnings": {
            "type": "array",
            "items": {"type": "string"}
          }
        }
      },
...
    }
  }
}